Most of the API's endpoints now use session tokens.
The API endpoints for admins, users, falls, and diary entries now use session tokens.

+++ web/public/api/diary-entries/delete.php
@@ -7,23 +7,30 @@
 		include_once '../models/DiaryEntry.php';
 
 		$api_key = isset($_GET['key']) ? $_GET['key'] : die(json_encode(array('message' => 'No API key provided.')));
-
-		$expected = ['entryID'];
-		$missing = [];
-
-		$database = new Database(false);
-		$db = $database->connect($api_key);
-
-		$diaryEntry = new DiaryEntry($db);
 		
 		$input = json_decode(file_get_contents('php://input'), true);
 
-		$diaryEntry->entryID = !empty($input['entryID']) ? $input['entryID'] : array_push($missing, 'entryID');
+		$expected = ['patientID', 'entryID'];
+		$missing = [];
 
-		if (empty($missing)) {
-			$diaryEntry->delete();
+		$patientID = !empty($input['patientID']) ? $input['patientID'] : array_push($missing, 'patientID');
+
+		$database = new Database();
+		
+		if ($database->verify(array('key' => $api_key, 'id' => $patientID))) {
+			$db = $database->connect();
+
+			$diaryEntry = new DiaryEntry($db);
+
+			$diaryEntry->entryID = !empty($input['entryID']) ? $input['entryID'] : array_push($missing, 'entryID');
+
+			if (empty($missing)) {
+				$diaryEntry->delete();
+			} else {
+				die(json_encode(array('expected' => $expected, 'missing' => $missing), JSON_PRETTY_PRINT));
+			}
 		} else {
-			die(json_encode(array('expected' => $expected, 'missing' => $missing), JSON_PRETTY_PRINT));
+			echo json_encode(array('message' => 'Invalid API key.'));
 		}
 	} else {
 		echo json_encode(array('message' => 'Wrong HTTP request method. Use DELETE instead.'));
